7-27 语音听书优化

+++ app/src/main/java/com/huli/page/model/bean/Timing.java
@@ -2,11 +2,11 @@ package com.huli.page.model.bean;
 
 public class Timing {
     public String name;
-    public int value;
+    public long value;
 
     public boolean isSelect;
 
-    public Timing(String name, int value) {
+    public Timing(String name, long value) {
         this.name = name;
         this.value = value;
     }
